% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_changeBGcolor.R
\name{change_bg_color}
\alias{change_bg_color}
\title{Function that change background color in NGL viewer}
\usage{
change_bg_color(input, output, is_pg2)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{is_pg2}{logical TRUE/FALSE. If TRUE, we calling to create result in page 2 (get latest results page).}
}
\description{
Function that change background color in NGL viewer
}
