% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_changeCAVcolor.R
\name{change_cav_color}
\alias{change_cav_color}
\title{Change color of cavities}
\usage{
change_cav_color(input, output, is_pg2)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{is_pg2}{logical TRUE/FALSE. If TRUE, we calling to create result in page 2 (get latest results page).}
}
\description{
Change color of cavities
}
