% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_changeSTRcolor.R
\name{change_str_color}
\alias{change_str_color}
\title{Function that changes the color of the biomolecular structure in NGL viewer}
\usage{
change_str_color(input, output, protein_col_list, is_pg2)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{protein_col_list}{list of previous selected colors}

\item{is_pg2}{logical TRUE/FALSE. If TRUE, we calling to create result in page 2 (get latest results page).}
}
\description{
Function that changes the color of the biomolecular structure in NGL viewer
}
