% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ui_chooseRunMode.R
\name{choose_run_mode}
\alias{choose_run_mode}
\title{Create radio buttons of the run mode in main Run KVFinder page}
\usage{
choose_run_mode()
}
\description{
Create radio buttons of the run mode in main Run KVFinder page
}
