% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_colorCAVhyd.R
\name{color_cavity_hyd}
\alias{color_cavity_hyd}
\title{Function that change background color in NGL viewer}
\usage{
color_cavity_hyd(input, output, is_pg2, cav_rep_list, result_pdb_list)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{is_pg2}{logical TRUE/FALSE. If TRUE, we calling to create result in page 2 (get latest results page).}

\item{result_pdb_list}{}
}
\description{
Function that change background color in NGL viewer
}
