% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ui_jumbotron.R
\name{create_jumbotron}
\alias{create_jumbotron}
\title{Function that creates the jumbotron of the main Run KVFinder page}
\usage{
create_jumbotron()
}
\description{
Function that creates the jumbotron of the main Run KVFinder page
}
