% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_boxCoord.R
\name{get_coord_box_mode}
\alias{get_coord_box_mode}
\title{Get box coordinates in box mode}
\usage{
get_coord_box_mode(pdb_input, box_residues)
}
\arguments{
\item{pdb_input}{Pdb path}

\item{box_residues}{List of target residues}
}
\description{
Get box coordinates in box mode
}
\examples{

}
