% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_interfaceRES.R
\name{interface_cav}
\alias{interface_cav}
\title{Select cavities to show}
\usage{
interface_cav(input, output, result_pdb_list, is_pg2)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{result_pdb_list}{a list returned by check function}

\item{is_pg2}{logical TRUE/FALSE. If TRUE, we calling to create result in page 2 (get latest results page).}
}
\description{
Select cavities to show
}
