% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kvParams.R
\name{kv_params}
\alias{kv_params}
\title{Create radio buttons of the run mode in main Run KVFinder page}
\usage{
kv_params(id, input_type)
}
\description{
Create radio buttons of the run mode in main Run KVFinder page
}
