% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_fetch.R
\name{process_fetch}
\alias{process_fetch}
\title{Function that creates the boxes and buttons of the run mode in fetch mode}
\usage{
process_fetch(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\description{
Function that creates the boxes and buttons of the run mode in fetch mode
}
