% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_fetchLigmode.R
\name{process_fetch_ligmode}
\alias{process_fetch_ligmode}
\title{Steps to process PDB obtained from fetch in ligand mode}
\usage{
process_fetch_ligmode(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\description{
Steps to process PDB obtained from fetch in ligand mode
}
