% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pdbProcess_.R
\name{report_nonstand}
\alias{report_nonstand}
\title{Get non-standard protein/nucleic residues}
\usage{
report_nonstand(pdb_input, show_modal)
}
\arguments{
\item{pdb_input}{Pdb path}

\item{show_modal}{}
}
\description{
Get non standard protein/nucleic residues from PDB
}
