% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_submitJob.R
\name{submit_job}
\alias{submit_job}
\title{Submit job}
\usage{
submit_job(input, output, run_mode_list, lig_name_list)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{run_mode_list}{list of latest run modes}

\item{lig_name_list}{list of latest ligand names}
}
\description{
Submit job
}
