% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_prepareSubmission.R
\name{submit_prepare}
\alias{submit_prepare}
\title{Prepare the input file for server submission}
\usage{
submit_prepare(
  pdb_path,
  ligand_path,
  whole_protein_mode,
  ligand_mode,
  box_mode,
  box_residues,
  probe_in,
  probe_out,
  volume_cutoff,
  removal_distance,
  padding,
  lig_cutoff
)
}
\description{
Prepare the input file for server submission
}
