% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_submitJob.R
\name{submit_job}
\alias{submit_job}
\title{Function that submits the KVFinder job to the server}
\usage{
submit_job(input, output, pdb_name_click_load, url_address, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{pdb_name_click_load}{variable of the number of clicks on load button}

\item{url_address}{url address to conexion}

\item{session}{}
}
\description{
Function that submits the KVFinder job to the server
}
